chat-integrations/slack/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
requests